backend/backend/requirements.txt
django>=5.0,<6.0
psycopg2-binary>=2.9
djangorestframework==3.15.1
django-cors-headers==4.3.1